abstract ExplosiveTool, make explosive mode persistent
