+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/tools/ExplosivePickaxe.java
@@ -1,13 +1,19 @@
 package io.github.thebusybiscuit.slimefun4.implementation.items.tools;
 
+import javax.annotation.Nonnull;
 import javax.annotation.ParametersAreNonnullByDefault;
 
+import org.bukkit.NamespacedKey;
 import org.bukkit.block.Block;
+import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 
+import io.github.bakedlibs.dough.protection.Interaction;
 import io.github.thebusybiscuit.slimefun4.api.items.ItemGroup;
 import io.github.thebusybiscuit.slimefun4.api.items.SlimefunItemStack;
 import io.github.thebusybiscuit.slimefun4.api.recipes.RecipeType;
+import io.github.thebusybiscuit.slimefun4.implementation.Slimefun;
+import io.github.thebusybiscuit.slimefun4.utils.tags.SlimefunTag;
 
 /**
  * The {@link ExplosivePickaxe} is a pickaxe which can destroy {@link Block}s
@@ -26,4 +32,23 @@ public class ExplosivePickaxe extends ExplosiveTool {
         super(itemGroup, item, recipeType, recipe);
     }
 
+    @Override
+    protected boolean canBreak(@Nonnull Player p, @Nonnull Block b) {
+        if (b.isEmpty() || b.isLiquid()) {
+            return false;
+        } else if (SlimefunTag.UNBREAKABLE_MATERIALS.isTagged(b.getType())) {
+            return false;
+        } else if (!b.getWorld().getWorldBorder().isInside(b.getLocation())) {
+            return false;
+        } else if (Slimefun.getIntegrations().isCustomBlock(b)) {
+            return false;
+        } else {
+            return Slimefun.getProtectionManager().hasPermission(p, b.getLocation(), Interaction.BREAK_BLOCK);
+        }
+    }
+
+    @Override
+    protected NamespacedKey getNamespacedKey() {
+        return new NamespacedKey(Slimefun.instance(), "explosive_pickaxe");
+    }
 }
