+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/tools/ExplosiveShovel.java
@@ -2,6 +2,7 @@ package io.github.thebusybiscuit.slimefun4.implementation.items.tools;
 
 import javax.annotation.ParametersAreNonnullByDefault;
 
+import org.bukkit.NamespacedKey;
 import org.bukkit.block.Block;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
@@ -35,4 +36,8 @@ public class ExplosiveShovel extends ExplosiveTool {
         return SlimefunTag.EXPLOSIVE_SHOVEL_BLOCKS.isTagged(b.getType()) && Slimefun.getProtectionManager().hasPermission(p, b.getLocation(), Interaction.BREAK_BLOCK);
     }
 
+    @Override
+    protected NamespacedKey getNamespacedKey() {
+        return new NamespacedKey(Slimefun.instance(), "explosive_shovel");
+    }
 }
